feat: move a common method into navigation service
